requirements.txt
asgiref==3.7.2
autopep8==2.0.4
certifi==2023.7.22
charset-normalizer==3.2.0
coreapi==2.3.3
coreschema==0.0.4
Django==4.2.5
django-mysql==4.11.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
idna==3.4
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.3
Pillow==10.0.1
pycodestyle==2.11.0
pytz==2023.3.post1
requests==2.31.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.8.0
uritemplate==4.1.1
urllib3==2.0.4
